requirements-dev.txt
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
autoflake==2.3.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.8
deepsearch-glm==1.0.0
dill==0.3.9
distro==1.9.0
docling==2.10.0
docling-core==2.9.0
docling-ibm-models==2.0.7
docling-parse==3.0.0
easyocr==1.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et_xmlfile==2.0.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
fastapi==0.112.1
filelock==3.16.1
filetype==1.2.0
fsspec==2024.10.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.5
idna==3.10
imageio==2.36.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain-core==0.3.21
langchain-text-splitters==0.3.2
langcodes==3.4.1
langdetect==1.0.9
langsmith==0.1.145
language_data==1.2.0
lazy_loader==0.4
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==3.0.2
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.17
murmurhash==1.0.10
networkx==3.4.2
ninja==1.11.1.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
openai==1.41.1
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
orjson==3.10.12
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pluggy==1.5.0
preshed==3.0.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
PyMuPDF==1.24.11
pymupdf4llm==0.0.17
PyPDF2==3.0.1
pypdfium2==4.30.0
pytest==8.3.3
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.3
rpds-py==0.22.3
rsa==4.9
Rtree==1.3.0
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
semchunk==2.2.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.38.6
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
thinc==8.2.5
tifffile==2024.9.20
tokenizers==0.21.0
tomli==2.1.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.5
transformers==4.47.0
triton==3.1.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.30.6
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
XlsxWriter==3.2.0
